cleanup

+++ src/modules/uniswap/CryptoWatcher.Modules.Uniswap/Abstractions/IUnichainEventFetcher.cs
@@ -1,3 +1,4 @@
+using System.Numerics;
 using CryptoWatcher.Modules.Uniswap.Entities;
 using CryptoWatcher.UniswapModule.Models;
 
@@ -6,7 +7,7 @@ namespace CryptoWatcher.Modules.Uniswap.Abstractions;
 public interface IUnichainEventFetcher
 {
     IAsyncEnumerable<List<LiquidityPoolPositionEvent>> FetchLiquidityPoolEvents(UniswapChainConfiguration chain,
-        ulong fromBlock,
-        ulong toBlock,
+        BigInteger fromBlock,
+        BigInteger toBlock,
         CancellationToken ct = default);
 }
